weather-pipeline/extract/requirements.txt
requests==2.31.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
python-dotenv==1.0.0
